+++ src/main/java/com/pahana/mapper/CustomerMapper.java
@@ -1,4 +1,4 @@
-package mapper;
+package com.pahana.mapper;
 
 import com.pahana.model.Customer;
 import com.pahana.dto.CustomerDTO;
